// src/psp/enums/TransactionStatus.ts
export enum TransactionStatus {
    INITIATED = 'initiated',
    SENT = 'sent',
    GET_LINK = 'get_link',
    GET_LINK_ERROR = 'get_link_error',
    GET_LINK_SUCCESS = 'get_link_success',
    OPENED = 'opened',
    PENDING = 'pending',
    AUTHENTICATION_START = 'authentication_start',
    AUTHENTICATION_APPROVED = 'authentication_approved',
    AUTHENTICATION_ERROR = 'authentication_error',
    RESULT_RECEIVED = 'result_received',
    PROCESSING_FAILED = 'processing_failed',
    REJECT = 'reject',
    ERROR = 'process_error',
    SUCCESS = 'success',
    COMPLETE = 'complete',
    CHECK_USER_EXISTS = 'check_user_start',
    USER_CREATE_EXTERNAL_START = 'user_create_external_start',
    USER_CREATE_EXTERNAL_END = 'user_create_external_end',
    USER_CREATE_EXTERNAL_FAILED = 'user_create_external_failed',
    GET_USER_KYC_START = 'get_user_kyc_start',
    GET_USER_KYC_END = 'get_user_kyc_end',
    USER_KYC_UPLOAD_START = 'user_kyc_upload_start',
    USER_KYC_UPLOAD_END = 'user_kyc_upload_end',
    USER_KYC_UPLOAD_FAILED = 'user_kyc_upload_failed',
    CHECK_USER_ERROR = 'check_user_failed',
}